Started moving away from arraylists for buses and shared stops
This also has the caveat of a smaller MapsActivity! Yay

Though it should be noted that by the end of this that I should go through and make sure that all the old artifacts of ArrayLists are removed (Such as converting from an array to an array list and back to an array, as that is super inefficient).