I want to start tackling the motherbug #13 .
I think I figured out a way to achieve this: Dont store the buses in the route, but rather within the activity itself, and then impliment some logic that executes the drawBuses() method (and have it dont asynchronously) whenever the selected routes change.

Also, remove additional copies of the selected arrays from the UpdateThread.